Develop

+++ TheSku/frmUser.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Linq;
 using System.Windows.Forms;
+using Telerik.WinControls;
 using TheSku.Data;
 
 namespace TheSku
@@ -12,6 +13,9 @@ namespace TheSku
         {
             this.dbContext = dbContext;
             InitializeComponent();
+            this.btnReload.Shortcuts.Add(new RadShortcut(Keys.Control, Keys.R));
+            this.btnDelete.Shortcuts.Add(new RadShortcut(Keys.Control, Keys.T));
+            this.btnCopyNameToClipboard.Shortcuts.Add(new RadShortcut(Keys.Alt, Keys.C));
             this.gvList.AutoGenerateColumns = false;
             this.ActiveControl = this.txtFirstName;
         }
@@ -100,10 +104,6 @@ namespace TheSku
             {
                 this.btnNew.PerformClick();
             }
-            if (e.Control && e.KeyCode == Keys.D)
-            {
-                this.btnDelete.PerformClick();
-            }
             if (e.KeyCode == Keys.F5)
             {
                 if (tabControl1.SelectedIndex == 0)
@@ -176,15 +176,18 @@ namespace TheSku
         {
             if (this.lblID.Text != "0")
             {
-                if (MessageBox.Show($"Are you sure you want to delete {this.lblID.Text}?", "Confirm", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                if (this.lblID.Text != Global.UserName)
                 {
-                    var supplier = dbContext.Users.Where(x => x.Name.Equals(this.lblID.Text)).FirstOrDefault();
-                    if (supplier is not null)
+                    if (MessageBox.Show($"Are you sure you want to delete {this.lblID.Text}?", "Confirm", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
                     {
-                        dbContext.Users.Remove(supplier);
-                        dbContext.SaveChanges();
-                        MessageBox.Show($"{this.lblID.Text} deleted successfully.", "Success", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                        this.btnNew.PerformClick();
+                        var supplier = dbContext.Users.Where(x => x.Name.Equals(this.lblID.Text)).FirstOrDefault();
+                        if (supplier is not null)
+                        {
+                            dbContext.Users.Remove(supplier);
+                            dbContext.SaveChanges();
+                            MessageBox.Show($"{this.lblID.Text} deleted successfully.", "Success", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                            this.btnNew.PerformClick();
+                        }
                     }
                 }
             }
@@ -218,6 +221,19 @@ namespace TheSku
             if (e.KeyCode == Keys.Enter && this.gvList.RowCount > 0)
             {
                 gvList_CellDoubleClick(sender, new DataGridViewCellEventArgs(this.gvList.CurrentCell.ColumnIndex, this.gvList.CurrentCell.RowIndex));
+            }
+        }
+
+        private void btnReload_Click(object sender, EventArgs e)
+        {
+            this.LoadData();
+        }
+
+        private void btnCopyNameToClipboard_Click(object sender, EventArgs e)
+        {
+            if (this.lblID.Text != "0")
+            {
+                Clipboard.SetText(this.lblID.Text);
             }
         }
     }
